requirements.txt
absl-py==1.4.0
backcall==0.2.0
bqplot==0.12.40
cachetools==5.3.1
cassandra-driver==3.28.0
#certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
config==0.5.1
cycler==0.11.0
DateTime==5.2
debugpy==1.7.0
decorator==5.1.1
diskcache==5.6.3
entrypoints==0.4
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
geomet==0.2.1.post1
google-auth==2.23.0
google-auth-oauthlib==0.4.6
grpcio==1.58.0
h5py==3.14.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.2
importlib-metadata==6.7.0
ipydatawidgets==4.3.5
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.10.1
ipyvuetify==1.8.10
ipywebrtc==0.6.0
ipywidgets==8.1.1
jedi==0.19.0
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==3.4.18.65
packaging==23.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pythreejs==2.4.2
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
SimpleITK==2.2.1
six==1.16.0
submitit==1.4.5
torchinfo==1.8.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.5
torch-summary==1.4.5
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
traittypes==0.2.1
typing_extensions==4.7.1
urllib3==1.26.16
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.9
yacs==0.1.8
zipp==3.15.0
zope.interface==6.0
